feat: separated module coder to extend file
