+++ caesar-cipher-cli/index.html
@@ -8,5 +8,6 @@
 </head>
 <body>
   <script src="./Caesar-cipher-tool.js"></script>
+    <script src="./coder.js"></script>
 </body>
 </html>
